// packages/src/infra/persistence/mysql/prisma/generate/models/PastGameRecord.ts
export type PastGameRecordModel = runtime.Types.Result.DefaultSelection<Prisma.$PastGameRecordPayload>
// ...
export type $PastGameRecordPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "PastGameRecord"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: number
    date: Date
    homeTeam: string
    awayTeam: string
    ballPark: string
    cancelled: boolean
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["pastGameRecord"]>
  composites: {}
}